Homme: fix ICE
In SCREAMv1 runs on summit, with new SurfaceCoupling, an ICE arises in
homme boundary exchange.
