+++ components/homme/src/preqx_kokkos/CMakeLists.txt
@@ -183,6 +183,7 @@ MACRO(PREQX_KOKKOS_SETUP)
     ${SRC_SHARE_DIR}/cxx/Tracers.cpp
     ${SRC_SHARE_DIR}/cxx/VerticalRemapManager.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/BoundaryExchange.cpp
+    ${SRC_SHARE_DIR}/cxx/mpi/BoundaryExchange_build_buffer_views_and_requests.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/Comm.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/Connectivity.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/MpiBuffersManager.cpp
