+++ components/homme/src/theta-l_kokkos/CMakeLists.txt
@@ -189,6 +189,7 @@ MACRO(THETAL_KOKKOS_SETUP)
     ${SRC_SHARE_DIR}/cxx/vertical_remap.cpp
     ${SRC_SHARE_DIR}/cxx/VerticalRemapManager.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/BoundaryExchange.cpp
+    ${SRC_SHARE_DIR}/cxx/mpi/BoundaryExchange_build_buffer_views_and_requests.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/Comm.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/Connectivity.cpp
     ${SRC_SHARE_DIR}/cxx/mpi/MpiBuffersManager.cpp
